requirements.txt
django>=3.0, <4.0

django-environ==0.4.5

django-ckeditor==6.1.0